moving functional_profiles to experimental base
