--- tests/api_data/functional_profiles.json
@@ -1,15 +1,0 @@
-[
-    {
-        "$uri": "/api/v1/functional_profiles/45a573fb7833449b",
-        "complete": true,
-        "created_at": "2015-09-26T06:16:39.794227-07:00",
-        "error_msg": "",
-        "job": {
-            "$ref": "/api/v1/jobs/c4f763556452402c"
-        },
-        "sample": {
-            "$ref": "/api/v1/samples/002f55a11aae4b8f"
-        },
-        "success": true
-    }
-]
